fix: #23 리스트 생성 테스트 코드로 변경

+++ matzipback/src/main/java/com/matzip/matzipback/matzipList/query/mapper/ListQueryMapper.java
@@ -1,6 +1,5 @@
 package com.matzip.matzipback.matzipList.query.mapper;
 
-import com.matzip.matzipback.matzipList.query.dto.ListCategoryDTO;
 import com.matzip.matzipback.matzipList.query.dto.ListSearchAllDTO;
 import org.apache.ibatis.annotations.Mapper;
 
@@ -11,9 +10,7 @@ public interface ListQueryMapper {
 
     List<ListSearchAllDTO> getListBox(long listUserSeq);
 
-
-    
-    Long getCountList(long listUserSeq); // 각 사용자의 리스트 개수 구하기
+    int getCountList(long listUserSeq); // 각 사용자의 리스트 개수 구하기
